% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fulltext.R
\name{fetch_fulltext}
\alias{fetch_fulltext}
\title{Get full-text articles from PMC Open Access Subset in an AWS S3 bucket}
\usage{
fetch_fulltext(id, vars = c("all", "section", "paragraph", "sentence", "text"))
}
\arguments{
\item{id}{a single PMCID, a vector of PMCIDs or a list of PMCIDs returned from \code{\link{fetch_pmcid}}}

\item{vars}{which variables you would like to keep in output dataframe, options
are \code{c("all", "section", "paragraph", "sentence", "text")}. Defaults to "all".}
}
\value{
a tidy \code{data.frame} containing each manuscript retrieved, tokenized by sentence.
}
\description{
Get full-text articles from PMC Open Access Subset in an AWS S3 bucket
}
\examples{
fetch_fulltext("PMC5465829")
fetch_fulltext("PMC5465829", vars = c("text"))
}
