% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_meta.R
\name{fetch_meta}
\alias{fetch_meta}
\title{Retrieve meta data for a paper given a PMCID or URL and/or specific meta data parameters and return a data.table}
\usage{
fetch_meta(id)
}
\arguments{
\item{id}{a single PMCID, a vector of PMCIDs or a list of PMCIDs returned from \code{\link{fetch_pmcid}}}
}
\value{
a \code{data.frame} containing all available meta data of an article in long format, given a PMID(s).
}
\description{
Retrieve meta data for a paper given a PMCID or URL and/or specific meta data parameters and return a data.table
}
\examples{
fetch_meta("17284678")
fetch_meta(17284678)
}
