% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pmcid.R
\name{fetch_pmcid}
\alias{fetch_pmcid}
\title{Convert PMIDs to PMCIDs}
\usage{
fetch_pmcid(...)
}
\arguments{
\item{...}{multiple strings holding possible drug names}
}
\value{
A named list of PMCIDs with the input PMIDs
}
\description{
Convert PMIDs to PMCIDs
}
\examples{
fetch_pmcid(19023454, "7016449")
}
