% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmcid.R
\name{get_pmcid}
\alias{get_pmcid}
\title{Convert PMIDs to PMCIDs}
\usage{
get_pmcid(...)
}
\arguments{
\item{...}{multiple strings holding possible drug names}
}
\value{
A named list of PMCIDs with the input PMIDs
}
\description{
Convert PMIDs to PMCIDs
}
\examples{
get_pmcid(19023454, "7016449")
}
